refactor: rename and move libelle functions
