+++ src/client/components/ui/Form/Combobox/ComboboxLocalisation/defaultLocalisation/buildUserInput.ts
@@ -1,19 +1,18 @@
-import { TypeLocalisation } from '~/server/localisations/domain/localisation';
-
-import {
-	formatLocalisationLibelle,
-} from '../ComboboxLocalisation';
 import {
 	DefaultLocalisation,
-} from './defaultLocalisation';
+} from '~/client/components/ui/Form/Combobox/ComboboxLocalisation/defaultLocalisation/defaultLocalisation';
+import {
+	formatLibelleLocalisation,
+} from '~/client/components/ui/Form/Combobox/ComboboxLocalisation/localisations/formatLibelleLocalisation';
+import { TypeLocalisation } from '~/server/localisations/domain/localisation';
 
 export function buildUserInput(defaultLocalisation?: DefaultLocalisation) {
 	if (!defaultLocalisation) return '';
 
 	if (defaultLocalisation.type === TypeLocalisation.DEPARTEMENT || defaultLocalisation.type === TypeLocalisation.REGION) {
-		return formatLocalisationLibelle(defaultLocalisation.nom, defaultLocalisation.code);
+		return formatLibelleLocalisation(defaultLocalisation.nom, defaultLocalisation.code);
 	} else if (defaultLocalisation.type === TypeLocalisation.COMMUNE) {
-		return formatLocalisationLibelle(defaultLocalisation.nom, defaultLocalisation.codePostal);
+		return formatLibelleLocalisation(defaultLocalisation.nom, defaultLocalisation.codePostal);
 	}
 	return '';
 }
