+++ src/client/components/ui/Form/Combobox/ComboboxLocalisation/localisations/formatLibelleLocalisation.ts
@@ -1,0 +1,3 @@
+export function formatLibelleLocalisation(nom: string, code: string) {
+	return `${nom} (${code})`;
+}
